Handle exceptions & add test for reading invalid frame
Also change assertions for GeneratorState to match the new single
status attribute (instead of status A, B and C) and leave assertions
for RuntimeState attributes commented while I don't figure out how to
deal with the new base::Time attributes.

The problem with the time attributes is that (probably due to type
conversion between C++ and ruby) the actual value is different from the
expected one. This wasn't a problem in the pure-C++ driver library
tests.